refactor(args): new module for args normalisation
